refactor(front): generate depth one record gql fields from record util
