+++ packages/twenty-front/src/modules/object-record/graphql/utils/generateDepthOneRecordGqlFields.ts
@@ -1,30 +1,14 @@
 import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
-import { isDefined } from 'twenty-shared';
 
+export type GenerateDepthOneRecordGqlFields = {
+  objectMetadataItem: ObjectMetadataItem;
+};
 export const generateDepthOneRecordGqlFields = ({
   objectMetadataItem,
-  record,
-}: {
-  objectMetadataItem: ObjectMetadataItem;
-  record?: Record<string, any>;
-}) => {
-  const gqlFieldsFromObjectMetadataItem = objectMetadataItem.fields.reduce<
-    Record<string, boolean>
-  >((acc, field) => {
+}: GenerateDepthOneRecordGqlFields) =>
+  objectMetadataItem.fields.reduce<Record<string, true>>((acc, field) => {
     return {
       ...acc,
       [field.name]: true,
     };
   }, {});
-
-  if (isDefined(record)) {
-    return Object.keys(gqlFieldsFromObjectMetadataItem).reduce((acc, key) => {
-      return {
-        ...acc,
-        [key]: Object.keys(record).includes(key),
-      };
-    }, gqlFieldsFromObjectMetadataItem);
-  }
-
-  return gqlFieldsFromObjectMetadataItem;
-};
